django_bookstore-main/django_bookstore-main/requirements.txt
asgiref==3.6.0
black==22.12.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.5
django-allauth==0.52.0
django-crispy-forms==1.14.0
gunicorn==20.1.0
idna==3.4
mypy-extensions==0.4.3
mysqlclient==2.1.1
oauthlib==3.2.2
pathspec==0.11.0
platformdirs==2.6.2
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.1
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.3.0
Pillow
python-dotenv
django
stripe
pyotp